Rename Codeblock component
